NEP-977 - Fix Dropdown values
Bug: When a dropdown is created as the 1+nth filter, it doesn't have
it's value initialized correctly. For some reason it was setting a
string as the initial value.

This bugfix patch should fix the behaviour of the dropdowns by checking
the props.value in the Component constructor. If this value is
undefined, or a blank string, the Component will initialize the Dropdown
element to select the first element of the list. Otherwise, it will
select the previously selected value.